Move some util functions around (#883)
